requirements.txt
aiogram==3.13.1
APScheduler==3.10.4
SQLAlchemy==2.0.23
openai==1.54.3
PyMuPDF==1.24.7
python-dotenv==1.0.1
dateparser==1.2.0
tzdata==2025.1
streamlit==1.32.2
